requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
spacy==3.7.2
python-docx==1.1.0
PyMuPDF==1.23.17